// types/jobs.interface.d.ts
export interface IJobs {
  id: number;
  OfficeType: string;
  JobType: string;
  BranchID: string;
  NumberOfJobs: string;
  Gender: string;
  Grade: string;
  MinAge: string;
  MaxAge: string;
  EXP: string;
  Salary: string;
  JobName: string;
  AddedBy: string;
  JobShift: string;
  JobNature: string;
  IntStartDate: string;
  AppLastDate: string;
  State: string;
  District: string;
  Taluk: string;
  Village: string;
  Status: string;
}